+++ src/sqlfmt/rules/common.py
@@ -29,8 +29,11 @@ SQL_COMMENT = group(
 )
 
 CREATE_FUNCTION = (
-    r"create(\s+or\s+replace)?(\s+temp(orary)?)?(\s+secure)?(\s+table)?"
+    r"create(\s+or\s+replace)?(\s+temp(orary)?)?(\s+secure)?"
+    r"(\s+external)?(\s+table)?"
     r"\s+function(\s+if\s+not\s+exists)?"
 )
+ALTER_DROP_FUNCTION = r"(alter|drop)\s+function(\s+if\s+exists)?"
+
 CREATE_WAREHOUSE = r"create(\s+or\s+replace)?\s+warehouse(\s+if\s+not\s+exists)?"
 ALTER_WAREHOUSE = r"alter\s+warehouse(\s+if\s+exists)?"