+++ src/sqlfmt/rules/function.py
@@ -2,7 +2,7 @@ from functools import partial
 
 from sqlfmt import actions
 from sqlfmt.rule import Rule
-from sqlfmt.rules.common import CREATE_FUNCTION, group
+from sqlfmt.rules.common import ALTER_DROP_FUNCTION, CREATE_FUNCTION, group
 from sqlfmt.rules.core import CORE
 from sqlfmt.token import TokenType
 
@@ -34,6 +34,7 @@ FUNCTION = [
         priority=1300,
         pattern=group(
             CREATE_FUNCTION,
+            ALTER_DROP_FUNCTION,
             r"language",
             r"transform",
             r"immutable",
@@ -50,17 +51,38 @@ FUNCTION = [
             r"cost",
             r"rows",
             r"support",
-            r"set",
             # snowflake
-            r"comment",
+            r"((un)?set\s+)?comment",
             r"imports",
             r"packages",
             r"handler",
             r"target_path",
             r"(not\s+)?null",
+            # snowflake external functions
+            r"((un)?set\s+)?"
+            + group(
+                r"api_integration",
+                r"headers",
+                r"context_headers",
+                r"max_batch_rows",
+                r"compression",
+                r"request_translator",
+                r"response_translator",
+            ),
             # bq
             r"options",
             r"remote\s+with\s+connection",
+            # ALTER
+            r"rename\s+to",
+            r"owner\s+to",
+            r"set\s+schema",
+            r"(no\s+)?depends\s+on\s+extension",
+            r"cascade",
+            r"restrict",
+            # alter snowflake
+            r"(un)?set\s+secure",
+            # pg catchall for set
+            r"(re)?set(\s+all)?",
         )
         + group(r"\W", r"$"),
         action=partial(actions.add_node_to_buffer, token_type=TokenType.UNTERM_KEYWORD),